Replace entries if already exist
For already existing entries, we can chose to either
1) replace them using INSERT OR REPLACE
But it will delete the record before inserting a new one if it
has to replace, which could be bad as we have foreign key cascades
or other delete triggers.
2) update them using INSERT .. UPDATE ..
But updating may leave certain fields to NULL value if not properly
handled.

Since any s3 PUT operations overwrites existing one, using
INSERT OR REPLACE for now.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>