Use mutex locking for thread synchronization
In a multi-threaded environment, below operations
need to be synchronized -

* Add/Delete/Find on objectmap
* Bind and Execute on saved prepared statements
* Multiple queries batched together for any operation.

To keep it simple, for now defined a single mutex variable
which will be used to protect all the above operations.
This mutex is locked for every operation done on base class
DBstore object (in ProcessOp()). If required, it can be made
further granular in future.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>